Mock dbus and gobject modules for readthedocs builds (#154)
The auto documentation of modules was not working on readthedocs
because the Bluezero modules were not loading because of missing
modules on the build server. This has been fixed by mocking the dbus
and gobject modules in the conf.py file.
Some tidy-up of docstrings. Mainly because __init__ strings are not
output