Reduce the number of examples
There were errors in the examples contained in the last release.
This is not good as examples are where people are likely to start with
the library.
As the library currently does not have automated testing of the
examples a more manageable number of examples are included.
There is one example for each role the library can support.
These will need to be tested manually before each release