dbus_tools: _get_dbus_path2 improvement
The error raising behavior of get_dbus path is not consistent, in fact:
In the case get_dbus_path is called with a correct adapter and
an erroneous device the function will silently return None.
In the case get_dbus_path is called with a correct device and an
erroneous adapter the function will raise a Value error because
_get_dbus_path is called with None on the second call.

Here we make the behavior uniform to allow better error management
where this function is used.
This change allows to uniformly return None when the dbus_path requested does not exist.